requirements.txt
chromadb
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

llama-index>=0.12.0
llama-index-llms-openai>=0.3.0
llama-index-embeddings-huggingface>=0.3.0
llama-index-vector-stores-postgres>=0.3.0
llama-index-vector-stores-chroma>=0.4.0
chromadb>=0.6.0
llama-index-readers-file>=0.3.0
llama-index-llms-openai-like>=0.3.0

psycopg[binary]>=3.2.0
pgvector>=0.3.0

celery[redis]>=5.4.0
redis>=5.2.0

pydantic-settings>=2.6.0
python-dotenv>=1.0.0


bcrypt==4.3.0                    # 固定这个版本，避免 5.0+ 的严格检查
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

slowapi>=0.1.9

unstructured>=0.16.0
pymupdf>=1.24.0

ragas>=0.2.0

pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.8.0
ruff>=0.6.0
mypy>=1.11.0
pre-commit>=4.0.0

sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.30.0
psycopg[binary]>=3.2.0